Restructure iter tests and simplify peek and next
